makefile: fixup demosoc/evalsoc compile error.

+++ Makefile
@@ -254,7 +254,7 @@ buildroot_busybox-menuconfig: $(buildroot_initramfs_wrkdir)/.config $(buildroot_
 $(buildroot_initramfs_sysroot_stamp): $(buildroot_initramfs_tar)
 	mkdir -p $(buildroot_initramfs_sysroot)
 	tar -xpf $< -C $(buildroot_initramfs_sysroot) --exclude ./dev --exclude ./usr/share/locale
-	cp -af $(confdir)/xy1000_net.ko $(buildroot_initramfs_sysroot)/usr/bin/
+	if [ -f $(confdir)/xy1000_net.ko ]; then cp -af $(confdir)/xy1000_net.ko $(buildroot_initramfs_sysroot)/usr/bin/; fi
 	touch $@
 
 .PHONY: initrd linux
@@ -431,7 +431,7 @@ prepare4m: buildroot_initramfs_sysroot
 	sed -i '/sbin\/getty/cconsole::respawn:/bin/sh' $(buildroot_initramfs_sysroot)/etc/inittab
 	#sed -i '/init\.d/d' $(buildroot_initramfs_sysroot)/etc/inittab
 	rm -rf $(buildroot_initramfs_sysroot)/lib/*
-	cp -af $(confdir)/xy1000_net.ko $(buildroot_initramfs_sysroot)/usr/bin/
+	if [ -f $(confdir)/xy1000_net.ko ]; then cp -af $(confdir)/xy1000_net.ko $(buildroot_initramfs_sysroot)/usr/bin/; fi
 	$(MAKE) CORE=$(CORE) cleanboot bootimages
 
 freeloader4m: prepare4m $(freeloader_elf)
